orders-service/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
sqlparse==0.5.3
gunicorn==22.0.0
psycopg2-binary
requests==2.32.4